apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "aurimasgriciunas", email = "griciunas.aurimas@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.57.0",
    "groq>=1.0.0",
    "langgraph>=1.0.5",
    "langsmith>=0.6.2",
    "openai>=2.15.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "qdrant-client>=1.16.2",
    "uvicorn>=0.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
shared = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/api"]

[tool.hatch.build.targets.sdist]
include = [
    "src/",
    "tests/",
    "README.md",
]
